updated the code based on reviews: remove comments;  relocate functions: operator_norm() and za_to_f64().
